Refactor & add additional tests + mocks
